Host/pyproject.toml
[project]
name = "cbor-host"
version = "0.1"
description = "CBOR Host"
readme = "README.md"
authors = [{name = "R. Keelan"}]
license = "Apache-2.0"
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "click",
    "pyserial"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
Run-Host = "cbor_host.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
